requirements.txt
absl-py===2.1.0
accelerate==0.29.2
aiohttp==3.9.5
bert_score==0.3.13
bitsandbytes==0.43.1
datasets==2.19.0
fastapi==0.111.0
fire==0.6.0
flash-attn==2.6.3
gradio==4.37.1
langdetect==1.0.9
llama_recipes==0.0.2
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
peft==0.10.0
pydantic==2.7.4
python-dotenv==1.0.1
rouge_score==0.1.2
tiktoken==0.7.0
torch_optimizer==0.3.0
transformers==4.44.2
triton==2.2.0
tqdm==4.66.2
trl==0.8.6
wandb==0.17.3
wheel==0.43.0
xformers==0.0.25